User Flags (State, Status, Quest) - user.hasflag, addflag, removeflag - death handling now uses userstate - blink uses userstate - spell chanting uses userstate - panelobject.canuse changed to virtual - clamp now accepts a double, because it's more modular

+++ Chaos/Client/Attributes.cs
@@ -50,8 +50,8 @@ namespace Chaos
         [JsonProperty]
         internal byte Ability;
 
-        internal uint MaximumHP => Utility.Clamp<uint>((int)(BaseHP + HPMod), 0, int.MaxValue);
-        internal uint MaximumMP => Utility.Clamp<uint>((int)(BaseMP + MPMod), 0, int.MaxValue);
+        internal uint MaximumHP => Utility.Clamp<uint>(BaseHP + HPMod, 0, int.MaxValue);
+        internal uint MaximumMP => Utility.Clamp<uint>(BaseMP + MPMod, 0, int.MaxValue);
         internal byte CurrentStr => Utility.Clamp<byte>(BaseStr + StrMod, 0, byte.MaxValue);
         internal byte CurrentInt => Utility.Clamp<byte>(BaseInt + IntMod, 0, byte.MaxValue);
         internal byte CurrentWis => Utility.Clamp<byte>(BaseWis + WisMod, 0, byte.MaxValue);
